feat: add line-helpers test suite to validate parsing and annotation functions
